Serialized saving and loading.
Renamed FileSystem to VirtualFileSystem for less confusing use in dialog.

Extracted save/loadShellState methods from Save/Load.executeCommand. Wrote
alternative versions that serialize the data, which required having both the
VirtualFileSystem hierarchy and IShellState be Serializable.

Ran into a snag: in loadShellSTate, we don't have access to method main's
local variables, so we couldn't quite restore the file system yet. We'll
have to redesign.

Fixed the unnecessary catch for FileNotFoundException. It will be caught as
an IOException.